backend/requirements.txt
# ModelForge-CV Requirements
# Python 3.9+

# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
peft>=0.5.0

# Training & Optimization
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Data Processing
pillow>=9.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0

# API & Backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
websockets>=11.0

# Task Queue & Workers
celery>=5.3.0
redis>=4.6.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# Storage & Cloud (optional)
# boto3>=1.28.0
# google-cloud-storage>=2.10.0

# Monitoring & Logging
loguru>=0.7.0
tqdm>=4.66.0

# Utilities
python-dotenv>=1.0.0